fis: [Enemy] 리소스 매니저 사용하게 수정

+++ Assets/Devs/HJD/Scripts/EnemyService.cs
@@ -59,7 +59,7 @@ public class EnemyService
         }
     }
 
-    private void SetEnemyData(EnemyViewModel vm,int enemyInstanceId, string enemyDataId, string name, int totalExp, string elementalType, int baseHp, float baseDamage,string PrefabAddress)
+    private void SetEnemyData(EnemyViewModel vm,int enemyInstanceId, string enemyDataId, string name, int totalExp, string elementalType, float baseHp, float baseDamage,string PrefabAddress)
     {
         vm.EnemyInstanceId = enemyInstanceId;
         vm.EnemyDataId = enemyDataId;
